Get event data before creating RSVP
In RSVP controller, first get the event using the event id from the request body.
Use that data to build an RSVP object.
Fill that object with the remaining data coming from the request body (the actual responses to the event questions).
Create a new RSVP.
Return the JSON from that RSVP.

(This code is heavily commented & very messy; but it's functional!)