Update Destroy on Events Controller
After learning that we mongo doesn't not automatically delete rsvps
when the event it's associated with is deleted, we updated the destroy
action on the events controller to fix this.

Now, when the delete request is made, first, the rsvps associated with
the given event id are deleted, and then the event itself is deleted.

Also updated the rsvp curl script to remove title, date and time since
we no longer need to pass those.