Use update with upsert instead of create for RSVPs
Use update with upsert inside of rsvps#create.

This prevents us from creating two RSVPs for the same user for the same event.
If an RSVP exists for that user for that event, update with upsert will overwrite (update) the existing RSVP.
If no RSVP exists, update with upsert will create a new RSVP.