Filter events
Return only events that user has not already RSVPed for from events#index

Get all events that do not belong to user. Examine each event's RSVPs. If none of the RSVPs have an owner that matches the current user, add that event to the filteredEvents array. Return the filteredEvents array.